+++ components/movementsensor/register/register.go
@@ -10,7 +10,6 @@ import (
 	_ "go.viam.com/rdk/components/movementsensor/gpsrtkserial"
 	_ "go.viam.com/rdk/components/movementsensor/imuvectornav"
 	_ "go.viam.com/rdk/components/movementsensor/imuwit"
-	_ "go.viam.com/rdk/components/movementsensor/imuwithwt905"
 	_ "go.viam.com/rdk/components/movementsensor/merged"
 	_ "go.viam.com/rdk/components/movementsensor/mpu6050"
 	_ "go.viam.com/rdk/components/movementsensor/replay"
